Workaround VC++Runtime deadlock bug by accessing thread HANDLE directly and calling TerminateThread().

+++ examples/containers/phonebills.cpp
@@ -71,7 +71,7 @@ struct ProduceBill
                    (last.event == 2 && e.event == 1));
 
             if (e.event == 2)            // call ended
-                sum += CT_PER_MIN * (e.timestamp - last.timestamp) / 60;
+                sum += CT_PER_MIN * (unsigned int)(e.timestamp - last.timestamp) / 60;
         }
         else if (last.from != -1)
         {
